Ensure `UnnamedType` is always assigned a type
For grammars with no unnamed types, `UnnamedType` will be `never` (which is correct). For all other grammars, the `never` in the union should be ignored.

An alternative which would require a significantly larger change would check for the absence of unnamed types and optionally outputting `never`, but this felt like a less intrusive change as an outside contributor