Ensure `constructor` methods include a return type
This is a workaround for the line above which converts TreeSitter’s class types to interfaces.

Note: it’s unclear to me what the purpose of that conversion actually is. I had assumed it would avoid name clashes with types generated by this project, but that doesn’t seem to be the case. There’s potential for other subtle differences between types allowed for classes versus interfaces, and it may make more sense to just eliminate the conversion instead. But again, this feels like a less intrusive change than that alternative.